Added new type of components - Unity-like ExecutableComponent Added fix for preventing multiple calls of Entity.Destroy()

+++ Source/BEntities/Entity.cs
@@ -100,6 +100,9 @@ namespace BEntities
 		/// </summary>
         public void Destroy()
         {
+            if(MarkedToBeRemoved)
+                return;
+
             Service.DestroyEntity(this);
 
 			// set servive reference as null for avoiding memory leak
